HSEARCH-3105 Separate the formatter in codec and formatter
