+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/formatter/impl/IntegerFieldFormatter.java
@@ -6,82 +6,17 @@
  */
 package org.hibernate.search.v6poc.backend.lucene.types.formatter.impl;
 
-import java.util.Objects;
-
-import org.apache.lucene.document.Document;
-import org.apache.lucene.document.IntPoint;
-import org.apache.lucene.document.NumericDocValuesField;
-import org.apache.lucene.document.StoredField;
-import org.apache.lucene.index.IndexableField;
-import org.hibernate.search.v6poc.backend.document.model.Sortable;
-import org.hibernate.search.v6poc.backend.document.model.Store;
-import org.hibernate.search.v6poc.backend.lucene.document.impl.LuceneDocumentBuilder;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFieldFormatter;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaObjectNode;
 
 public final class IntegerFieldFormatter implements LuceneFieldFormatter<Integer> {
 
-	private final Store store;
+	public static final IntegerFieldFormatter INSTANCE = new IntegerFieldFormatter();
 
-	private final Sortable sortable;
-
-	public IntegerFieldFormatter(Store store, Sortable sortable) {
-		this.store = store;
-		this.sortable = sortable;
-	}
-
-	@Override
-	public void addFields(LuceneDocumentBuilder documentBuilder, LuceneIndexSchemaObjectNode parentNode, String fieldName, Integer value) {
-		if ( value == null ) {
-			return;
-		}
-
-		if ( Store.YES.equals( store ) ) {
-			documentBuilder.addField( parentNode, new StoredField( fieldName, value ) );
-		}
-
-		if ( Sortable.YES.equals( sortable ) ) {
-			documentBuilder.addField( parentNode, new NumericDocValuesField( fieldName, value.longValue() ) );
-		}
-
-		documentBuilder.addField( parentNode, new IntPoint( fieldName, value ) );
-	}
-
-	@Override
-	public Integer parse(Document document, String fieldName) {
-		IndexableField field = document.getField( fieldName );
-
-		if ( field == null ) {
-			return null;
-		}
-
-		return (Integer) field.numericValue();
+	private IntegerFieldFormatter() {
 	}
 
 	@Override
 	public Object format(Object value) {
 		return value;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if ( this == obj ) {
-			return true;
-		}
-		if ( obj == null ) {
-			return false;
-		}
-		if ( IntegerFieldFormatter.class != obj.getClass() ) {
-			return false;
-		}
-
-		IntegerFieldFormatter other = (IntegerFieldFormatter) obj;
-
-		return Objects.equals( store, other.store ) && Objects.equals( sortable, other.sortable );
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash( store, sortable );
 	}
 }
