+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/types/formatter/impl/LocalDateFieldFormatter.java
@@ -6,109 +6,19 @@
  */
 package org.hibernate.search.v6poc.backend.lucene.types.formatter.impl;
 
-import static java.time.temporal.ChronoField.DAY_OF_MONTH;
-import static java.time.temporal.ChronoField.MONTH_OF_YEAR;
-import static java.time.temporal.ChronoField.YEAR;
-
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeFormatterBuilder;
-import java.time.format.ResolverStyle;
-import java.time.format.SignStyle;
-import java.util.Locale;
-import java.util.Objects;
 
-import org.apache.lucene.document.Document;
-import org.apache.lucene.document.LongPoint;
-import org.apache.lucene.document.NumericDocValuesField;
-import org.apache.lucene.document.StoredField;
-import org.apache.lucene.index.IndexableField;
-import org.hibernate.search.v6poc.backend.document.model.Sortable;
-import org.hibernate.search.v6poc.backend.document.model.Store;
-import org.hibernate.search.v6poc.backend.lucene.document.impl.LuceneDocumentBuilder;
 import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneFieldFormatter;
-import org.hibernate.search.v6poc.backend.lucene.document.model.impl.LuceneIndexSchemaObjectNode;
 
 public final class LocalDateFieldFormatter implements LuceneFieldFormatter<LocalDate> {
 
-	private static final DateTimeFormatter FORMATTER = new DateTimeFormatterBuilder()
-			.appendValue( YEAR, 4, 9, SignStyle.EXCEEDS_PAD )
-			.appendLiteral( '-' )
-			.appendValue( MONTH_OF_YEAR, 2 )
-			.appendLiteral( '-' )
-			.appendValue( DAY_OF_MONTH, 2 )
-			.toFormatter( Locale.ROOT )
-			.withResolverStyle( ResolverStyle.STRICT );
+	public static final LocalDateFieldFormatter INSTANCE = new LocalDateFieldFormatter();
 
-	private final Store store;
-
-	private final Sortable sortable;
-
-	public LocalDateFieldFormatter(Store store, Sortable sortable) {
-		this.store = store;
-		this.sortable = sortable;
-	}
-
-	@Override
-	public void addFields(LuceneDocumentBuilder documentBuilder, LuceneIndexSchemaObjectNode parentNode, String fieldName, LocalDate value) {
-		if ( value == null ) {
-			return;
-		}
-
-		if ( Store.YES.equals( store ) ) {
-			documentBuilder.addField( parentNode, new StoredField( fieldName, FORMATTER.format( value ) ) );
-		}
-
-		long valueToEpochDay = value.toEpochDay();
-
-		if ( Sortable.YES.equals( sortable ) ) {
-			documentBuilder.addField( parentNode, new NumericDocValuesField( fieldName, valueToEpochDay ) );
-		}
-
-		documentBuilder.addField( parentNode, new LongPoint( fieldName, valueToEpochDay ) );
+	private LocalDateFieldFormatter() {
 	}
 
 	@Override
 	public Object format(Object value) {
 		return ((LocalDate) value).toEpochDay();
-	}
-
-	@Override
-	public LocalDate parse(Document document, String fieldName) {
-		IndexableField field = document.getField( fieldName );
-
-		if ( field == null ) {
-			return null;
-		}
-
-		String value = field.stringValue();
-
-		if ( value == null ) {
-			return null;
-		}
-
-		return LocalDate.parse( value, FORMATTER );
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if ( this == obj ) {
-			return true;
-		}
-		if ( obj == null ) {
-			return false;
-		}
-		if ( LocalDateFieldFormatter.class != obj.getClass() ) {
-			return false;
-		}
-
-		LocalDateFieldFormatter other = (LocalDateFieldFormatter) obj;
-
-		return Objects.equals( store, other.store ) && Objects.equals( sortable, other.sortable );
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash( store, sortable );
 	}
 }
